Move three core-related tests to core module (#2731)
* Move two core-related api tests to core

* Move suppression test to core as suppression was moved to internal and will become core

* Clean up api dependencies

* Remove unused properties
