--- detekt-api/src/test/kotlin/io/gitlab/arturbosch/detekt/api/Case.kt
@@ -1,13 +1,0 @@
-package io.gitlab.arturbosch.detekt.api
-
-import io.github.detekt.test.utils.resourceAsPath
-import java.nio.file.Path
-
-enum class Case(val file: String) {
-    FilteredClass("FilteredClass.kt"),
-    SuppressedObject("SuppressedObject.kt"),
-    SuppressedByAllObject("SuppressedByAllObject.kt"),
-    SuppressedElements("SuppressedElements.kt");
-
-    fun path(): Path = resourceAsPath(file)
-}
