--- detekt-api/src/test/resources/FilteredClass.kt
@@ -1,3 +1,0 @@
-@Suppress("unused")
-class FilteredClass {
-}
