+++ detekt-core/src/test/resources/deactivated_ruleset.yml
@@ -1,0 +1,2 @@
+comments:
+  active: false
