+++ detekt-core/src/test/resources/pathFilters/multi-rule-without-excludes.yml
