+++ detekt-core/src/test/resources/suppression/SuppressedByAllObject.kt
