+++ detekt-core/src/test/resources/suppression/SuppressedElements.kt
